pyproject.toml
[project]
name = "iris-shield"
version = "0.2.0"
description = "Anti-VLM privacy protection: make your photos unreadable to AI vision models while looking normal to humans."
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }

dependencies = [
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.40.0",
    "open-clip-torch>=2.24.0",
    "ultralytics>=8.3.0",
    "easyocr>=1.7.0",
    "Pillow>=10.2.0",
    "numpy>=1.26.0,<2.0.0",
    "opencv-python-headless>=4.9.0",
    "streamlit>=1.40.0",
    "plotly>=5.18.0",
    "loguru>=0.7.2",
    "huggingface_hub>=0.20.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.3.0", "pytest>=8.0.0"]

[tool.ruff]
line-length = 100
target-version = "py310"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu124"
name = "pytorch-cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
